webapi/tct-alarm-tizen-tests/alarm/AlarmManager_getAll.html

<!DOCTYPE html>

<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

*Redistributions of works must retain the original copyright notice, this list
of conditions and the following disclaimer.
*Redistributions in binary form must reproduce the original copyright notice,
this list of conditions and the following disclaimer in the documentation
and/or other materials provided with the distribution.
*Neither the name of Intel Corporation nor the names of its contributors
may be used to endorse or promote products derived from this work without
specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jenny Cao <jenny.q.cao@intel.com>
        Mariusz Polasinski <m.polasinski@samsung.com>

Revision history:

-->

<html>

<head>
<title>AlarmManager_getAll</title>
<meta charset="utf-8">
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
<script type="text/javascript" src="support/alarm_common.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: AlarmManager_getAll
//==== LABEL Check using getAll method in AlarmManager interface to get all alarms
//==== SPEC Tizen Web API:Application:Alarm:AlarmManager:getAll M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/alarm.html
//==== PRIORITY P1
//==== TEST_CRITERIA MR MNA

test(function () {
    add_result_callback(function () {
        try {
            tizen.alarm.removeAll();
        } catch (err) {
            // do nothing in case removeAll throw an exception
        }
    });

    var i, absAlarm, relAlarm, alarmAll = [];
    cleanAlarms();
    absAlarm = createAbsAlarm();
    tizen.alarm.add(absAlarm, APPLICATION_ID);
    relAlarm = createRelAlarm();
    tizen.alarm.add(relAlarm, APPLICATION_ID);
    alarmAll = tizen.alarm.getAll();
    assert_equals(alarmAll.length, 2, "number of added alarms should be 2");
    for (i = 0; i < alarmAll.length; i++) {
        if (alarmAll[i] instanceof tizen.AlarmAbsolute) {
            assert_equals(alarmAll[i].id, absAlarm.id, "object is not the same which was added");
        } else if (alarmAll[i] instanceof tizen.AlarmRelative) {
            assert_equals(alarmAll[i].id, relAlarm.id, "object is not the same which was added");
        } else {
            assert_unreached("object was not get properly");
        }
    }
}, document.title);

</script>
</body>
</html>
